+++ ui/metalinkcreator/metalinker.h
@@ -140,7 +140,7 @@ class Metaurl
          * the priority of the urls, 1 is highest priority, 255 lowest
          * default is 0 as in not set and thus is ranked even behind 255
          */
-        uint priority;
+        quint16 priority;
 
         /**
          * Optional the name of a file that should be get of that metaurl
@@ -174,7 +174,7 @@ class Url
          * the priority of the urls, 1 is highest priority, 255 lowest
          * default is 0 as in not set and thus is ranked even behind 255
          */
-        uint priority;
+        quint16 priority;
 
         /**
          * the location of the server eg. "uk"
@@ -319,7 +319,7 @@ class Metalink
         Files files;
 
         static const QString KGET_DESCRIPTION;
-        static const uint MAX_PRIORITY; //maximum pirority a Metalink 4.0 Url or Metaurl can have, not to be mixed up with the highest priority
+        static const quint16 MAX_PRIORITY; //maximum pirority a Metalink 4.0 Url or Metaurl can have, not to be mixed up with the highest priority
 };
 
 /**
@@ -363,7 +363,7 @@ class Metalink_v3
 
     private:
         Metalink m_metalink;
-        static const uint MAX_PREFERENCE;
+        static const quint16 MAX_PREFERENCE;
 };
 
 /**
